Chapter02/storagedrive/requirements.txt
google-cloud-storage==1.6.0
google-cloud-datastore==1.4.0
